feat: streamedInbetween dropdown is a modal on mobile
